requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
numpy==1.24.3
pytest==7.3.1
python-dotenv==1.0.0
pymysql==1.0.3
cryptography==3.4.8